+++ src/gmxlib/nrama.c
@@ -269,7 +269,8 @@ static void get_dih_props(t_xrama *xr,t_idef *idef)
     
     key.ai[1]=ia[2];
     key.ai[2]=ia[3];
-    if ((dd = bsearch(&key,xr->dih,xr->ndih,sizeof(key),d_comp)) != NULL) {
+    if ((dd = bsearch(&key,xr->dih,xr->ndih,(size_t)sizeof(key),d_comp))
+	!= NULL) {
       dd->mult=idef->iparams[ft].pdihs.mult;
       dd->phi0=idef->iparams[ft].pdihs.phiA;
     }
